Add an advanced features area
Add a new expander with some "advanced" features mainly used for
administrators.

Add a new entry that allows to lookup participants by name that do not
appear in the shown list. This feature is useful in environments where
the user/group database is in a LDAP (or other NIS) that limits the
number of results in a query. Note that this code requires stuff only
available as of gtkmm 2.16 (which I think is pretty common now).

Move the checkbox of "showing system participants" in that expander as
well.

Also translate into English some identifiers still in Catalan.